Delete the many-to-many relationship from car to customer. Fix the primary key from Car ID to licencePlate. Fix the controller and service layer to a new primary key. Add a new relationship between car and customer. Fix DTO car and Customer.

+++ src/main/java/nl/novi/Eindopdracht/Service/CustomerAccountService.java
@@ -1,26 +1,29 @@
 package nl.novi.Eindopdracht.Service;
 
+import lombok.AllArgsConstructor;
+import nl.novi.Eindopdracht.Exceptions.CarNotFoundException;
 import nl.novi.Eindopdracht.Exceptions.RecordNotFoundException;
+import nl.novi.Eindopdracht.Models.Data.Car;
 import nl.novi.Eindopdracht.Models.Data.CustomerAccount;
 import nl.novi.Eindopdracht.Repository.CarOwnerRepository;
+import nl.novi.Eindopdracht.Repository.CarRepository;
 import nl.novi.Eindopdracht.Repository.CustomerAccountRepository;
 import nl.novi.Eindopdracht.dto.input.CustomerAccountDto;
+import nl.novi.Eindopdracht.dto.output.CarOutputDto;
 import nl.novi.Eindopdracht.dto.output.CustomerAccountOutputDto;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
+import java.util.*;
+@AllArgsConstructor
 @Service
 public class CustomerAccountService {
 
 
     private final CustomerAccountRepository cARepos;
 
-    public CustomerAccountService(CustomerAccountRepository cARepos,CarOwnerRepository carOwnerRepos) {
-        this.cARepos = cARepos;
-    }
+    private final CarRepository carRepository;
+
+
 
 
     public Long createCostumer(CustomerAccountDto.CustomerAccountAllDto cADto) {
@@ -33,7 +36,7 @@ public class CustomerAccountService {
     public List<CustomerAccountOutputDto> getAllCustomers() {
         List<CustomerAccountOutputDto> collection = new ArrayList<>();
         List<CustomerAccount> list = cARepos.findAll();
-        for (CustomerAccount account: list){
+        for (CustomerAccount account : list) {
             collection.add(accountToDTo1(account));
         }
         return collection;
@@ -41,10 +44,9 @@ public class CustomerAccountService {
 
     public CustomerAccountOutputDto getCustomerById(long id) {
         Optional<CustomerAccount> account = cARepos.findById(id);
-        if (!account.isPresent()){
+        if (!account.isPresent()) {
             throw new RecordNotFoundException("cannot find customer please enter a new id ");
-        }
-        else {
+        } else {
             CustomerAccount a = account.get();
             return accountToDTo1(a);
         }
@@ -52,10 +54,9 @@ public class CustomerAccountService {
 
     public CustomerAccountOutputDto.CustomerNameOutputDto getCustomerByLastName(String customerLastName) {
         Optional<List<CustomerAccount>> optionalAccounts = cARepos.findByLastName(customerLastName);
-        if(!optionalAccounts.isPresent()){
+        if (!optionalAccounts.isPresent()) {
             throw new RecordNotFoundException("cannot find customer please enter a anther name");
-        }
-        else {
+        } else {
             CustomerAccount a = (CustomerAccount) optionalAccounts.get();
             return accountToDTo2(a);
         }
@@ -63,40 +64,64 @@ public class CustomerAccountService {
     }
 
     public CustomerAccountOutputDto.CustomerFinanceOutputDto getBillingAddressById(long id) {
-        Optional<CustomerAccount>accountOptional = cARepos.findById(id);
-        if (!accountOptional.isPresent()){
-            throw  new RecordNotFoundException("cannot find the billing address so please enter a anther id");
-        }
-        else{
+        Optional<CustomerAccount> accountOptional = cARepos.findById(id);
+        if (!accountOptional.isPresent()) {
+            throw new RecordNotFoundException("cannot find the billing address so please enter a anther id");
+        } else {
             CustomerAccount a = accountOptional.get();
             return accountToDTo3(a);
         }
     }
 
-    public CustomerAccountDto.CustomerNameDto updateCustomerName(long id, String firstName, String LastName) {
-            Optional<CustomerAccount> accountOptional = cARepos.findById(id);
+    public Collection<CarOutputDto> getCarByCustomerName(String customerName) {
+        Collection<CarOutputDto> dtoCollection = new HashSet<>();
+        Optional<CustomerAccount> account = cARepos.findAllByCustomerName(customerName);
+        for (CustomerAccount:
+             account) {
+            Car car = car.get();
+            CarOutputDto dto = new CarOutputDto();
 
-            if (!accountOptional.isPresent()){
-                throw new RecordNotFoundException("Can not find "+ firstName + LastName + "so please enter a anther id ");
-            }
-            else {
-                CustomerAccount name = accountOptional.get();
-                name.setFirstName(firstName);
-                name.setLastName(LastName);
-                cARepos.save(name);
-                return null;
-            }
+            dto.setId(car.getId());
+            dto.setBrand(car.getBrand());
+            dto.setModel(car.getModel());
+            dto.setYearOfBuild(car.getYearOfBuild());
+            dto.setColor(car.getColor());
+            dto.setLicensePlate(car.getLicensePlate());
+            dto.setMileage(car.getMileage());
+            dto.setEngineType(car.getEngineType());
+            dto.setBody(car.getBody());
+            dto.setTransmission(car.getTransmission());
+            dto.setFuel(car.getFuel());
+
+            dtoCollection.add(dto);
+
+        }
+        return dtoCollection;
+    }
+
+
+    public CustomerAccountDto.CustomerNameDto updateCustomerName(long id, String firstName, String LastName) {
+        Optional<CustomerAccount> accountOptional = cARepos.findById(id);
+
+        if (!accountOptional.isPresent()) {
+            throw new RecordNotFoundException("Can not find " + firstName + LastName + "so please enter a anther id ");
+        } else {
+            CustomerAccount name = accountOptional.get();
+            name.setFirstName(firstName);
+            name.setLastName(LastName);
+            cARepos.save(name);
+            return null;
+        }
 
 
     }
 
     public CustomerAccountDto.CustomerFinanceDto updateFinance(long id, String billingAdress, String bankAccountNumber) {
-        Optional<CustomerAccount> accountOptional =cARepos.findById(id);
-        if(!accountOptional.isPresent()){
+        Optional<CustomerAccount> accountOptional = cARepos.findById(id);
+        if (accountOptional.isEmpty()) {
             throw new RecordNotFoundException("cannot find the files, please give me anther id");
-        }
-        else {
-            CustomerAccount a =accountOptional.get();
+        } else {
+            CustomerAccount a = accountOptional.get();
             a.setBankAccountNumber(bankAccountNumber);
             a.setBillingAddress(billingAdress);
             cARepos.save(a);
@@ -104,13 +129,22 @@ public class CustomerAccountService {
         }
     }
 
-    public void deleteCustomerById(long id) {
-        cARepos.deleteById(id);
+    public String deleteCustomerById(long id) {
+        if (!cARepos.existsById(id)) {
+            throw new CarNotFoundException("Car with id:" + id + "is not found");
+        } else {
+            long count = cARepos.count();
+            cARepos.deleteById(id);
+            return "You delete" + count + "in the id" + id;
+        }
     }
 
-    public void deleteAllCustomers() {
+    public String deleteAllCustomers() {
+        long count = cARepos.count();
         cARepos.deleteAll();
+        return "You deleted " + count + "cars";
     }
+
 
     public CustomerAccountOutputDto accountToDTo1(CustomerAccount account) {
         CustomerAccountOutputDto dto = new CustomerAccountOutputDto();
@@ -119,14 +153,14 @@ public class CustomerAccountService {
         dto.customerName = account.getCustomerName();
         dto.firstName = account.getFirstName();
         dto.lastName = account.getLastName();
-        dto.customerNumber= account.getCustomerNumber();
         dto.address = account.getAddress();
-        dto.phoneNumber =account.getPhoneNumber();
+        dto.phoneNumber = account.getPhoneNumber();
         dto.billingAddress = account.getBillingAddress();
-        dto.bankAccountNumber =account.getBankAccountNumber();
+        dto.bankAccountNumber = account.getBankAccountNumber();
 
         return dto;
     }
+
     public CustomerAccountOutputDto.CustomerNameOutputDto accountToDTo2(CustomerAccount account) {
         CustomerAccountOutputDto.CustomerNameOutputDto dto = new CustomerAccountOutputDto.CustomerNameOutputDto();
 
@@ -137,6 +171,7 @@ public class CustomerAccountService {
 
         return dto;
     }
+
     public CustomerAccountOutputDto.CustomerFinanceOutputDto accountToDTo3(CustomerAccount account) {
         CustomerAccountOutputDto.CustomerFinanceOutputDto dto = new CustomerAccountOutputDto.CustomerFinanceOutputDto();
 
@@ -146,17 +181,16 @@ public class CustomerAccountService {
         return dto;
     }
 
-    public CustomerAccount DtoToAccount(CustomerAccountDto.CustomerAccountAllDto accountDto){
-        CustomerAccount account =new CustomerAccount();
+    public CustomerAccount DtoToAccount(CustomerAccountDto.CustomerAccountAllDto accountDto) {
+        CustomerAccount account = new CustomerAccount();
 
-            account.setCustomerName(accountDto.customerName);
-            account.setFirstName(accountDto.firstName);
-            account.setLastName(accountDto.lastName);
-            account.setCustomerNumber(accountDto.costumerNumber);
-            account.setAddress(accountDto.address);
-            account.setPhoneNumber(accountDto.phoneNumber);
-            account.setBillingAddress(accountDto.billingAddress);
-            account.setBankAccountNumber(accountDto.bankAccountNumber);
+        account.setCustomerName(accountDto.customerName);
+        account.setFirstName(accountDto.firstName);
+        account.setLastName(accountDto.lastName);
+        account.setAddress(accountDto.address);
+        account.setPhoneNumber(accountDto.phoneNumber);
+        account.setBillingAddress(accountDto.billingAddress);
+        account.setBankAccountNumber(accountDto.bankAccountNumber);
 
 
         return account;
@@ -164,4 +198,8 @@ public class CustomerAccountService {
 
 
 
+    public CustomerAccountOutputDto getAccountByLicensePlate(String licensePlate) {
+    }
+
 }
+
